# backend/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserResponseType": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"editor"
						]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"role",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserInput": {
				"properties": {
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"editor"
						]
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Definições de entrada para criação e atualização.\r\nUpdateUserInput é definido localmente para permitir atualizações parciais (opcionais)."
			},
			"UpdateUserInput": {
				"properties": {
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"editor"
						]
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Token JWT no formato 'Bearer {token}'"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "Backend para o TCC Usina Guara",
		"contact": {}
	},
	"paths": {
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponseType"
								}
							}
						}
					},
					"409": {
						"description": "Conflict - Email já está em uso"
					}
				},
				"description": "Cria um novo usuário no sistema.",
				"summary": "Registrar novo usuário",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Dados obrigatórios do usuário (email, password, etc)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserInput",
								"description": "Dados obrigatórios do usuário (email, password, etc)"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserResponseType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Lista todos os usuários cadastrados.",
				"summary": "Listar usuários",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponseType"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Lista todos os usuários cadastrados.",
				"summary": "Listar usuários",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponseType"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Atualiza informações de um usuário existente.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID do usuário a ser modificado",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Campos opcionais para atualização",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserInput",
								"description": "Campos opcionais para atualização"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Remove um usuário do sistema.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID do usuário a ser excluído",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}